refactor: 예약시스템 메서드명 수정
