+++ src/main/java/com/wanted/preonboarding/ticket/application/TicketSeller.java
@@ -1,6 +1,7 @@
 package com.wanted.preonboarding.ticket.application;
 
 import com.wanted.preonboarding.ticket.domain.dto.*;
+import com.wanted.preonboarding.ticket.domain.dto.request.CreateReservationRequest;
 import com.wanted.preonboarding.ticket.domain.entity.Performance;
 import com.wanted.preonboarding.ticket.domain.entity.Reservation;
 import com.wanted.preonboarding.ticket.aop.dto.BaseResDto;
@@ -33,14 +34,14 @@ public class TicketSeller {
             .toList();
     }
 
-    public BaseResDto getPerformanceInfoDetail(ReserveInfo reserveInfo) {
+    public BaseResDto getPerformanceInfoDetail(CreateReservationRequest createReservationRequest) {
         log.info("getPerformanceInfoDetail");
-        Performance performance = getPerformance(reserveInfo.getPerformanceId(), reserveInfo.getRound());
-        Reservation reservation = getReservation(reserveInfo, performance);
+        Performance performance = getPerformance(createReservationRequest.getPerformanceId(), createReservationRequest.getRound());
+        Reservation reservation = getReservation(createReservationRequest, performance);
         return ResponseReserveQueryDto.of(performance, reservation);
     }
 
-    private Reservation getReservation(ReserveInfo reserveInfo, Performance performance) {
+    private Reservation getReservation(CreateReservationRequest reserveInfo, Performance performance) {
         return reservationRepository.findByPerformanceIdAndRoundAndLineAndSeat(
                 performance.getId(), performance.getRound(), reserveInfo.getLine(), reserveInfo.getSeat())
                 .orElseThrow(() -> new ServiceException(ResultCode.NOT_FOUND));
@@ -52,25 +53,25 @@ public class TicketSeller {
     }
 
     @Transactional
-    public void reserve(ReserveInfo reserveInfo) {
-        Performance performance = getPerformance(reserveInfo.getPerformanceId(), reserveInfo.getRound());
+    public void createReservation(CreateReservationRequest createReservationRequest) {
+        Performance performance = getPerformance(createReservationRequest.getPerformanceId(), createReservationRequest.getRound());
         String enableReserve = performance.getIsReserve();
         if (enableReserve.equalsIgnoreCase(ReserveStatus.ENABLE.getValue())) {
             // 1. 결제
             int price = performance.getPrice();
-            reserveInfo.setAmount(reserveInfo.getAmount() - price);
+            createReservationRequest.setAmount(createReservationRequest.getAmount() - price);
 
             // 2. 예매 된 좌석인지 확인
-            checkIsReserved(reserveInfo);
+            checkIsReserved(createReservationRequest);
 
             // 3. 예매 진행
-            reservationRepository.save(Reservation.from(reserveInfo));
+            reservationRepository.save(Reservation.from(createReservationRequest));
         } else {
             throw new ServiceException(ResultCode.RESERVE_NOT_VALID);
         }
     }
 
-    private void checkIsReserved(ReserveInfo reserveInfo) {
+    private void checkIsReserved(CreateReservationRequest reserveInfo) {
         reservationRepository.findByPerformanceIdAndRoundAndLineAndSeat(
                                         reserveInfo.getPerformanceId(),
                                         reserveInfo.getRound(),
