+++ src/main/java/com/wanted/preonboarding/ticket/domain/dto/ReserveSystemDto.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Builder
 public class ReserveSystemDto {
 //    private PerformanceInfo performanceInfo;
-//    private ReserveInfo reserveInfo;
+//    private CreateReservationRequest reserveInfo;
     private ResponseReserveQueryDto responseQueryDto;
 }
