+++ src/main/java/com/wanted/preonboarding/ticket/domain/dto/request/CreateReservationRequest.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.ticket.domain.dto;
+package com.wanted.preonboarding.ticket.domain.dto.request;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -8,15 +8,14 @@ import java.util.UUID;
 
 @Getter @Setter
 @Builder
-public class ReserveInfo {
-    // 공연 및 전시 정보 + 예약자 정보
-    private UUID performanceId; // 예약을 원하는 공연 또는 전시회ID
+public class CreateReservationRequest {
     private String reservationName; // 고객의 이름
     private String reservationPhoneNumber; // 휴대 전화
-    private String reservationStatus; // 예약; 취소;
     private long amount; // 결제 가능한 금액(잔고)
-    private int gate;
+    private UUID performanceId; // 예약을 원하는 공연 또는 전시회ID
     private int round; // 회차
+    private int gate; // 좌석 정보
     private char line; // 좌석 정보
     private int seat; // 좌석 정보
+    private String reservationStatus; // 예약; 취소;
 }
